improve csv extraction and separate resource

+++ inst/tools/phase1-workflow.sh
@@ -126,7 +126,7 @@ if [ -x "$(command -v rec2csv)" ]; then
   fi
 
   TASKSCSV="rec.tasks.csv.gz"
-  rec2csv tasks.rec | sed 's/"//g' | gzip -c > $TASKSCSV
+  (rec2csv tasks.rec | sed 's/"//g' | gzip -c > $TASKSCSV &)
   PAPIFILE="papi.rec"
   if [ -f "$PAPIFILE" ] && [ -s "$PAPIFILE" ]; then
     PAPICSV="rec.papi.csv.gz"
@@ -137,9 +137,6 @@ else
   echo "WARN: The library recutils is required to read the data.rec and tasks.rec files, skipping this step."
 fi
 
-if [ -z "$STARVZ_KEEP" ]; then
-  rm -f activity.data distrib.data trace.html tasks.rec papi.rec data.rec trace.rec
-fi
 echo "Convert from paje.sorted.trace to paje.csv"
 date "+%a %d %b %Y %H:%M:%S %Z"
 
@@ -158,40 +155,58 @@ fi
 if [ -z "$STARVZ_KEEP" ]; then
   rm -f paje.sorted.trace.gz
 fi
+echo "Wait"
+date "+%a %d %b %Y %H:%M:%S %Z"
+wait
+
+if [ -z "$STARVZ_KEEP" ]; then
+  rm -f activity.data distrib.data trace.html tasks.rec papi.rec data.rec trace.rec
+fi
+
 echo "Get states, links and variables in CSV"
 date "+%a %d %b %Y %H:%M:%S %Z"
 
-PAJE_MEMORY_STATE=paje.memory_state.csv.gz
-echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_MEMORY_STATE
-zgrep -e "Memory Node State" paje.csv.gz | gzip -c >> $PAJE_MEMORY_STATE
-
-PAJE_COMM_STATE=paje.comm_state.csv.gz
-echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_COMM_STATE
-zgrep -e "Communication Thread State" paje.csv.gz | gzip -c >> $PAJE_COMM_STATE
-
-PAJE_WORKER_STATE=paje.worker_state.csv.gz
-echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value, Size, Params, Footprint, Tag, JobId, SubmitOrder, GFlop, X, Y, Iteration, Subiteration" | gzip -c > $PAJE_WORKER_STATE
-if [ -n "$STARVZ_EXCLUDE_TASKS" ]; then
-  zgrep -e "Worker State" paje.csv.gz | grep -E -v $(echo $STARVZ_EXCLUDE_TASKS | sed "s/,/|/g" | sed "s/ //g" ) | gzip -c >> $PAJE_WORKER_STATE
-else
-  zgrep -e "Worker State" paje.csv.gz | sed -e 's/\(State\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,,,,,,,,,/g' | gzip -c >> $PAJE_WORKER_STATE
+if [ ! -f starvz_fast_csv_split ] && [ -x "$(command -v gcc)" ]; then
+  gcc -O2 $DIR/starvz_csv.c -o $DIR/starvz_fast_csv_split -lz -fopenmp
 fi
 
-PAJE_OTHER_STATE=paje.other_state.csv.gz
-echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_OTHER_STATE
-zgrep -E "^State" paje.csv.gz | grep -E -v "(Memory Node State|Communication Thread State|Worker State)" | gzip -c >> $PAJE_OTHER_STATE
+if [ -f $DIR/starvz_fast_csv_split ]; then
+  $DIR/starvz_fast_csv_split
+else
 
-PAJEVARIABLE=paje.variable.csv.gz
-echo "Nature, ResourceId, Type, Start, End, Duration, Value" | gzip -c > $PAJEVARIABLE
-zgrep -E "^Variable" paje.csv.gz | gzip -c >> $PAJEVARIABLE
+  PAJE_MEMORY_STATE=paje.memory_state.csv.gz
+  echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_MEMORY_STATE
+  zgrep -e "Memory Node State" paje.csv.gz | gzip -c >> $PAJE_MEMORY_STATE
 
-PAJELINK=paje.link.csv.gz
-echo "Nature, Container, Type, Start, End, Duration, Size, Origin, Dest, Key, Tag, MPIType, Priority, Handle" | gzip -c > $PAJELINK
-zgrep -E "^Link" paje.csv.gz | sed -e 's/\(Link\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,/g' | gzip -c >> $PAJELINK
+  PAJE_COMM_STATE=paje.comm_state.csv.gz
+  echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_COMM_STATE
+  zgrep -e "Communication Thread State" paje.csv.gz | gzip -c >> $PAJE_COMM_STATE
 
-PAJEEVENT=paje.events.csv.gz
-echo "Nature, Container, Type, Start, Value, Handle, Info, Size, Tid, Src" | gzip -c > $PAJEEVENT
-zgrep -E "^Event" paje.csv.gz | sed -e 's/\(Event\)\(,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,,,/g' | sed -e 's/\(Event\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,/g' | gzip -c >> $PAJEEVENT
+  PAJE_WORKER_STATE=paje.worker_state.csv.gz
+  echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value, Size, Params, Footprint, Tag, JobId, SubmitOrder, GFlop, X, Y, Iteration, Subiteration" | gzip -c > $PAJE_WORKER_STATE
+  if [ -n "$STARVZ_EXCLUDE_TASKS" ]; then
+    zgrep -e "Worker State" paje.csv.gz | grep -E -v $(echo $STARVZ_EXCLUDE_TASKS | sed "s/,/|/g" | sed "s/ //g" ) | gzip -c >> $PAJE_WORKER_STATE
+  else
+    zgrep -e "Worker State" paje.csv.gz | sed -e 's/\(State\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,,,,,,,,,/g' | gzip -c >> $PAJE_WORKER_STATE
+  fi
+
+  PAJE_OTHER_STATE=paje.other_state.csv.gz
+  echo "Nature, ResourceId, Type, Start, End, Duration, Depth, Value" | gzip -c > $PAJE_OTHER_STATE
+  zgrep -E "^State" paje.csv.gz | grep -E -v "(Memory Node State|Communication Thread State|Worker State)" | gzip -c >> $PAJE_OTHER_STATE
+
+  PAJEVARIABLE=paje.variable.csv.gz
+  echo "Nature, ResourceId, Type, Start, End, Duration, Value" | gzip -c > $PAJEVARIABLE
+  zgrep -E "^Variable" paje.csv.gz | gzip -c >> $PAJEVARIABLE
+
+  PAJELINK=paje.link.csv.gz
+  echo "Nature, Container, Type, Start, End, Duration, Size, Origin, Dest, Key, Tag, MPIType, Priority, Handle" | gzip -c > $PAJELINK
+  zgrep -E "^Link" paje.csv.gz | sed -e 's/\(Link\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,/g' | gzip -c >> $PAJELINK
+
+  PAJEEVENT=paje.events.csv.gz
+  echo "Nature, Container, Type, Start, Value, Handle, Info, Size, Tid, Src" | gzip -c > $PAJEEVENT
+  zgrep -E "^Event" paje.csv.gz | sed -e 's/\(Event\)\(,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,,,,,/g' | sed -e 's/\(Event\)\(,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*,[^,]*$\)/\1\2,/g' | gzip -c >> $PAJEEVENT
+
+fi
 
 if [ -z "$STARVZ_KEEP" ]; then
   rm -f paje.csv.gz
